requirements.txt
# Label Studio 项目依赖

# 基础依赖
label-studio>=1.9.0

# 自动标注依赖
opencv-python>=4.8.0
requests>=2.31.0
anthropic>=0.18.0
openai>=1.12.0

# 可选：本地模型支持
# ultralytics>=8.0.0  # YOLOv8
# torch>=2.0.0        # PyTorch
# torchvision>=0.15.0

# 工具库
Pillow>=10.0.0
numpy>=1.24.0

